agp-dashboard-web/pyproject.toml
[tool.poetry]
name = "agp-dashboard-web"
version = "0.1.0"
description = "A web dashboard for the AGP project."
authors = ["Alejandro Fernandez <noemisanalex@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.12, <4.0"
flask = ">=3.1.2,<4.0.0"
flask-login = ">=0.6.3,<0.7.0"
flask-socketio = ">=5.5.1,<6.0.0"
requests = ">=2.32.5,<3.0.0"
eventlet = ">=0.40.3,<0.41.0"
flask-sqlalchemy = ">=3.1.1,<4.0.0"
websocket-client = ">=1.8.0,<2.0.0"
apscheduler = ">=3.11.0,<4.0.0"
gunicorn = ">=23.0.0,<24.0.0"
tzdata = ">=2025.2,<2026.0"
flask-migrate = ">=4.1.0,<5.0.0"
flask-babel = "==4.0.0"
flask-compress = ">=1.18,<2.0"
pyjwt = ">=2.10.1,<3.0.0"
structlog = ">=25.4.0,<26.0.0"
pyyaml = ">=6.0.1,<7.0.0"
flask-swagger-ui = ">=5.21.0,<6.0.0"
flask-apscheduler = ">=1.13.1,<2.0.0"
google-auth-oauthlib = ">=1.2.2,<2.0.0"
google-generativeai = ">=0.5.4,<1.0.0"
cryptography = ">=42.0.5,<43.0.0"
dateparser = ">=1.2.0,<2.0.0"


[tool.poetry.group.dev.dependencies]
ruff = ">=0.5.0,<0.6.0"
mypy = ">=1.10.0,<2.0.0"
mkdocs-material = ">=9.0.0,<10.0.0"
pytest = "^8.4.2"
pytest-cov = "^7.0.0"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"